Kotlin task

+++ src/main/kotlin/NewStudent.kt
@@ -1,4 +1,4 @@
-data class NewStudent(val name: String, val  mark: Int, val exams: Int) { // в дата классе всегда используется val
+data class NewStudent(val name: String, val  mark: Float, val exams: Int) { // в дата классе всегда используется val
     val subject = 12 // не будет считываться так как находится не в конструкторе
 
     override fun toString(): String {
